conformance/functionalityTesting/basics/createContext.html

<!--
 Copyright (C) 2013 Samsung Electronics Corporation. All rights reserved.

 Redistribution and use in source and binary forms, with or without
 modification, are permitted provided the following conditions
 are met:

 1.  Redistributions of source code must retain the above copyright
     notice, this list of conditions and the following disclaimer.

 2.  Redistributions in binary form must reproduce the above copyright
     notice, this list of conditions and the following disclaimer in the
     documentation and/or other materials provided with the distribution.

 THIS SOFTWARE IS PROVIDED BY SAMSUNG ELECTRONICS CORPORATION AND ITS
 CONTRIBUTORS "AS IS", AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING
 BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
 FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL SAMSUNG
 ELECTRONICS CORPORATION OR ITS CONTRIBUTORS BE LIABLE FOR ANY DIRECT,
 INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES(INCLUDING
 BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
 DATA, OR PROFITS, OR BUSINESS INTERRUPTION), HOWEVER CAUSED AND ON ANY THEORY
 OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT(INCLUDING
 NEGLIGENCE OR OTHERWISE ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
 EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 -->
<html>
<head>
<meta charset="utf-8">
    <title>createContext conformance test.</title>
    <link rel="stylesheet" href="../../../resources/js-test-style.css"/>
    <script src="../../../resources/js-test-pre.js"></script>
    <script src="../../../resources/cl_utils.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script>
try {
    debug("CreateContext functionality test case.");

    var wtu = WebCLTestUtils;
    var webCLPlatform = wtu.getPlatform();
    var webCLDevices = wtu.getDevices(webCLPlatform);
    var webCLContext, deviceType, contextDevices, contextProperty;

    var createContext = function(platform, devices, type) {
        try {
            webCLContext = wtu.createContext(platform, devices, type);

            contextDevices = webCLContext.getInfo(webcl.CONTEXT_DEVICES);
            contextProperty = webCLContext.getInfo(webcl.CONTEXT_PROPERTIES);

            if (typeof(platform) != 'undefined' && platform != null) {
                if (contextProperty.platform !== platform)
                    return "CONTEXT_PROPERTIES::platform !== WebCLContext::platform.";

                for (i = 0; i < contextDevices.length; i++)
                    if (contextDevices[i].getInfo(webcl.DEVICE_PLATFORM) !== platform)
                        return "CONTEXT_PROPERTIES::platform !== WebCLContext::device.platform.";
            }

            if (typeof(devices) != 'undefined' && devices != null) {
                for (i = 0; i < contextDevices.length; i++)
                    if (contextDevices[i] !== devices[i] ||
                            contextDevices[i].getInfo(webcl.DEVICE_PLATFORM) !== contextProperty.platform)
                        return "CONTEXT_PROPERTIES::CONTEXT_DEVICES !== WebCLContext::devices.";
            }

            if (typeof(type) != 'undefined' && type != null) {
                // ALL and default cannot be checked for. They are just alias to other types.
                if (type == webcl.DEVICE_TYPE_ALL || type == webcl.DEVICE_TYPE_DEFAULT)
                    return true;

                if (contextProperty.deviceType != type)
                    return "CONTEXT_PROPERTIES::type != type of WebCLContext::devices.";

                for (i = 0; i < contextDevices.length; i++)
                    if (contextDevices[i].getInfo(webcl.DEVICE_TYPE) != type)
                        return "CONTEXT_PROPERTIES::devices.type != type of WebCLContext::devices.";
            }
            return true;

        } catch(e) {
            e.description = "Exception thrown while creating and validating context : " + e.name;
        }
    }

    shouldBeTrue("createContext()");
    shouldBeTrue("createContext(webCLPlatform)");
    shouldBeTrue("createContext(webCLPlatform, webCLDevices)");
    shouldBeTrue("createContext(null, webCLDevices)");
    shouldBeTrue("createContext(null, null, webcl.DEVICE_TYPE_CPU)");
    shouldBeTrue("createContext(null, null, webcl.DEVICE_TYPE_GPU)");
    shouldBeTrue("createContext(null, null, webcl.DEVICE_TYPE_ACCELERATOR)");
    shouldBeTrue("createContext(null, null, webcl.DEVICE_TYPE_ALL)");
    shouldBeTrue("createContext(null, null, webcl.DEVICE_TYPE_DEFAULT)");

} catch(e) {
    testFailed(e.description);
}
successfullyParsed = true;
</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
</html>
